day 10...needed help on part 2
got through part1 on my own, but with a verbose match that I think can be condensed down...but not too unhappy with it.

part 2 I stared at for a while and then went for help on. Even with help I struggled through it. Some of that was reading comprehension (the animal can live in junk pieces :shrugging:. and some was me missing that the type of piece the start was is important. Tests were passing but my input had it such that the start piece was one that shouldn't have toggled the boarder crossing

oh, it didn't help that I got past part 1 with some bogus paths in my
match from X back to Start (thankfully tests caught that one but took me
way too long since I had false confidence in that)